modified FPGA charge sense to be compatible with BQ24125
simply remove the pin, as the passive pull-ups do the right thing.
The charge sense changed between DVT and PVT, so this change should
be "safe" across both generations.

Note that lots of files are checked in because windows sucks and
changed a whole bunch of file permissions, I don't know why it
decided to do that.